+++ src/VirtoCommerce.ExperienceApiModule.DigitalCatalog/Breadcrumbs/CategoryBreadcrumb.cs
@@ -1,0 +1,14 @@
+using VirtoCommerce.CatalogModule.Core.Model;
+
+namespace VirtoCommerce.XDigitalCatalog.Breadcrumbs
+{
+    public class CategoryBreadcrumb : Breadcrumb
+    {
+        public CategoryBreadcrumb(Category category) : base(nameof(Category))
+        {
+            Category = category;
+        }
+
+        public Category Category { get; private set; }
+    }
+}
