+++ src/VirtoCommerce.ExperienceApiModule.DigitalCatalog/Extensions/ProductExtension.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using VirtoCommerce.CatalogModule.Core.Model;
 using VirtoCommerce.StoreModule.Core.Model;
+using VirtoCommerce.XDigitalCatalog.Breadcrumbs;
 
 namespace VirtoCommerce.XDigitalCatalog.Extensions
 {
